pyproject.toml
[project]
name = "sqm"
version = "0.1.0"
description = "Complex Langevin simulation for Bose-Hubbard model on 3D lattice"
authors = [
    { name = "terauchi hiroshi", email = "terauchi.hiroshi@cloud-ace.jp" }
]
dependencies = [
    "matplotlib>=3.9.0",
    "seaborn>=0.13.2",
    "numpy>=1.26.4",
    "click>=8.1.0",
    "pyyaml>=6.0.0",
]
readme = "README.md"
requires-python = ">= 3.12"

[project.scripts]
sqm = "sqm.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.4.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    "types-PyYAML>=6.0.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/sqm"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["F401", "F811"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
check_untyped_defs = true
no_implicit_optional = true

[[tool.mypy.overrides]]
module = [
    "matplotlib.*",
    "seaborn.*",
]
ignore_missing_imports = true
